app/requirements-docker.txt
Django>=3.1.4,<3.2.0
djangorestframework>=3.12.2,<3.13.0
django-cors-headers>=3.5.0,<3.6.0

psycopg2-binary>=2.8.6,<2.9.0
